Added client certificate Common Name check
The Common Name written in the client certificate used to connect to
remote JMX with SSL and client authentication is now checked to
correspond to the username provided in (username, password) credentials
that are used for authentication and JMX permissions attribution.
Added a test demonstrating that the connection with a client certificate
with a Common Name different from the username in credentials is
refused.